tts-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
torch>=2.1.0
torchaudio>=2.1.0
coqui-tts>=0.26.2
numpy>=1.24.3
librosa>=0.10.1
soundfile>=0.12.1
python-dotenv>=1.0.0
aiofiles>=23.2.1